# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeRepositoryInput {
  /** @minLength 1 */
  repositoryUrl: string;
}

export interface AskRepositoryInput {
  /** @minLength 1 */
  repositoryUrl: string;
  /** @minLength 1 */
  question: string;
}

export type RepositoryFileKind = typeof RepositoryFileKind[keyof typeof RepositoryFileKind];


export const RepositoryFileKind = {
  entry: 'entry',
  config: 'config',
  module: 'module',
  docs: 'docs',
  test: 'test',
} as const;

export interface RepositoryFile {
  path: string;
  kind: RepositoryFileKind;
  description: string;
}

export interface SelectedRepositoryFile {
  path: string;
  score: number;
  includedChars: number;
  truncated: boolean;
}

export interface RepositoryAnalysis {
  repositoryUrl: string;
  owner: string;
  repositoryName: string;
  description: string;
  defaultBranch: string;
  overview: string;
  language: string;
  stars: number;
  entryPoints: RepositoryFile[];
  keyFiles: RepositoryFile[];
  architecture: string[];
  analyzedAt: string;
  selectedFileCount: number;
  selectedTotalChars: number;
  selectedFiles: SelectedRepositoryFile[];
}

export interface RepositoryAnswer {
  answer: string;
  sources: string[];
}

export interface ErrorResponse {
  error: string;
}