# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: repositories
    description: Mock repository analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analyze:
    post:
      operationId: analyzeRepository
      tags: [repositories]
      summary: Analyze a public GitHub repository
      description: Returns realistic mock repository understanding data for a valid public GitHub URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRepositoryInput"
      responses:
        "200":
          description: Repository analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryAnalysis"
        "400":
          description: Invalid repository URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ask:
    post:
      operationId: askRepository
      tags: [repositories]
      summary: Ask a question about the repository
      description: Returns a mock source-cited answer for a repository question.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRepositoryInput"
      responses:
        "200":
          description: Answer with source filenames
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryAnswer"
        "400":
          description: Invalid question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnalyzeRepositoryInput:
      type: object
      properties:
        repositoryUrl:
          type: string
          minLength: 1
      required:
        - repositoryUrl
    AskRepositoryInput:
      type: object
      properties:
        repositoryUrl:
          type: string
          minLength: 1
        question:
          type: string
          minLength: 1
      required:
        - repositoryUrl
        - question
    RepositoryFile:
      type: object
      properties:
        path:
          type: string
        kind:
          type: string
          enum: [entry, config, module, docs, test]
        description:
          type: string
      required:
        - path
        - kind
        - description
    RepositoryAnalysis:
      type: object
      properties:
        repositoryUrl:
          type: string
        owner:
          type: string
        repositoryName:
          type: string
        description:
          type: string
        defaultBranch:
          type: string
        overview:
          type: string
        language:
          type: string
        stars:
          type: integer
        entryPoints:
          type: array
          items:
            $ref: "#/components/schemas/RepositoryFile"
        keyFiles:
          type: array
          items:
            $ref: "#/components/schemas/RepositoryFile"
        architecture:
          type: array
          items:
            type: string
        analyzedAt:
          type: string
        selectedFileCount:
          type: integer
        selectedTotalChars:
          type: integer
        selectedFiles:
          type: array
          items:
            $ref: "#/components/schemas/SelectedRepositoryFile"
      required:
        - repositoryUrl
        - owner
        - repositoryName
        - description
        - defaultBranch
        - overview
        - language
        - stars
        - entryPoints
        - keyFiles
        - architecture
        - analyzedAt
        - selectedFileCount
        - selectedTotalChars
        - selectedFiles
    SelectedRepositoryFile:
      type: object
      properties:
        path:
          type: string
        score:
          type: integer
        includedChars:
          type: integer
        truncated:
          type: boolean
      required:
        - path
        - score
        - includedChars
        - truncated
    RepositoryAnswer:
      type: object
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            type: string
      required:
        - answer
        - sources
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error