# lib/api-zod/src/generated/types/repositoryAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RepositoryFile } from './repositoryFile';
import type { SelectedRepositoryFile } from './selectedRepositoryFile';

export interface RepositoryAnalysis {
  repositoryUrl: string;
  owner: string;
  repositoryName: string;
  description: string;
  defaultBranch: string;
  overview: string;
  language: string;
  stars: number;
  entryPoints: RepositoryFile[];
  keyFiles: RepositoryFile[];
  architecture: string[];
  analyzedAt: string;
  selectedFileCount: number;
  selectedTotalChars: number;
  selectedFiles: SelectedRepositoryFile[];
}